+++ src/modules/archlinux.rs
@@ -1,118 +1,131 @@
-// =================== Types ==========================
+use crate::app::AppData;
+
+use super::Sendable;
+use anyhow::Context;
 use serde::Deserialize;
 
+const ARCH_PKG_SEARCH_API: &str = "https://www.archlinux.org/packages/search/json";
+
 #[derive(Deserialize)]
-pub struct SearchResponse {
+pub struct ArchLinuxSearchResponse {
     valid: bool,
-    results: Vec<PackageInfo>,
+    results: Vec<ArchLinuxPkgInfo>,
+}
+
+impl ArchLinuxSearchResponse {
+    #[inline]
+    pub fn is_valid(&self) -> bool {
+        self.valid
+    }
+
+    #[inline]
+    pub fn results(&self) -> &[ArchLinuxPkgInfo] {
+        &self.results
+    }
+
+    #[inline]
+    pub fn is_empty(&self) -> bool {
+        self.results.is_empty()
+    }
 }
 
 #[derive(Deserialize, Debug)]
-pub struct PackageInfo {
-    pkgname: String,
-    repo: String,
-    pkgver: String,
-    pkgrel: String,
-    pkgdesc: String,
-    url: String,
-    installed_size: u32,
-    last_update: String,
+pub struct ArchLinuxPkgInfo {
+    pub pkgname: String,
+    pub repo: String,
+    pub pkgver: String,
+    pub pkgrel: String,
+    pub pkgdesc: String,
+    pub url: String,
+    pub installed_size: u32,
+    pub last_update: String,
 }
 
-impl std::fmt::Display for PackageInfo {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "Name: {}
-Repo: {}
-Version: {}-{}
-Description: {}
-Upstream: {}
-Installed Size: {}
-Last Update: {}
-",
-            self.pkgname,
-            self.repo,
-            self.pkgver,
-            self.pkgrel,
-            self.pkgdesc,
-            self.url,
-            self.installed_size,
-            self.last_update
-        )
+pub async fn fetch_pkg_info(data: AppData, pkg: &str) -> anyhow::Result<Sendable> {
+    let url = reqwest::Url::parse_with_params(ARCH_PKG_SEARCH_API, &[("name", pkg)])
+        .with_context(|| format!("{pkg} is a invalid params"))?;
+
+    let resp: ArchLinuxSearchResponse = data.requester.to_t(url).await?;
+    if !resp.is_valid() {
+        anyhow::bail!("invalid request!")
     }
+
+    let pkg = resp
+        .results()
+        .iter()
+        .next()
+        .ok_or_else(|| anyhow::anyhow!("no result found for {pkg}"))?;
+
+    let display = format!(
+        "Name: {} \
+         Repo: {} \
+         Version: {}-{} \
+         Description: {} \
+         Upstream: {} \
+         Installed Size: {} \
+         Last Update: {} \
+       ",
+        pkg.pkgname,
+        pkg.repo,
+        pkg.pkgver,
+        pkg.pkgrel,
+        pkg.pkgdesc,
+        pkg.url,
+        pkg.installed_size,
+        pkg.last_update
+    );
+
+    Ok(Sendable::builder().text(display).build())
 }
 
-/// Types that implement ArchLinuxPacman trait should have two method:
-/// `pacman -Si` and `pacman -Ss`.
-#[async_trait::async_trait]
-pub trait ArchLinuxPkgProvider {
-    /// Customize search output type
-    type SearchOutput;
-    /// Act like `pacman -Ss`, and with `max` limit the output items
-    async fn search_pkg(&self, pkg: &str, max: usize) -> Self::SearchOutput;
-    /// Act like `pacman -Si`
-    async fn get_pkg_info(&self, pkg: &str) -> anyhow::Result<PackageInfo>;
-}
+pub async fn fetch_pkg_list(data: AppData, pkg: &str, max: usize) -> anyhow::Result<Sendable> {
+    let query_by = |typ: &str| -> anyhow::Result<reqwest::Url> {
+        reqwest::Url::parse_with_params(ARCH_PKG_SEARCH_API, &[(typ, pkg)])
+            .with_context(|| format!("{pkg} is a invalid params"))
+    };
 
-use crate::maid::Fetcher;
-use anyhow::Context;
+    let req = &data.requester;
+    let (exact_match, fuzzy_match) = tokio::join! {
+         req.to_t::<ArchLinuxSearchResponse>(query_by("name")?),
+         req.to_t::<ArchLinuxSearchResponse>(query_by("q")?),
+    };
 
-const SEARCH_BASE_URL: &str = "https://www.archlinux.org/packages/search/json";
+    let (exact_match, fuzzy_match) = (exact_match?, fuzzy_match?);
 
-#[async_trait::async_trait]
-impl ArchLinuxPkgProvider for Fetcher {
-    async fn get_pkg_info(&self, pkg: &str) -> anyhow::Result<PackageInfo> {
-        let url = reqwest::Url::parse_with_params(SEARCH_BASE_URL, &[("name", pkg)])
-            .with_context(|| format!("{pkg} is a invalid params"))?;
-
-        let resp = self.to_t::<SearchResponse>(url).await?;
-        if !resp.valid {
-            anyhow::bail!("invalid request!")
-        }
-
-        resp.results
-            .into_iter()
-            .next()
-            .ok_or_else(|| anyhow::anyhow!("no result found for {pkg}"))
+    if !exact_match.is_valid() || !fuzzy_match.is_valid() {
+        anyhow::bail!("invalid request!")
     }
 
-    type SearchOutput = anyhow::Result<(Option<String>, Vec<String>)>;
+    // Format characters: 19
+    //   + Max repo size (Community): 9
+    //   + General pkgname size: 10
+    //   + General pkgdesc size: 50
+    //
+    // * Entry counts: max
+    const MAYBE_CHAR_SIZE: usize = 19 + 9 + 10 + 50;
+    let fuzzy_pkg_size = fuzzy_match.results().len();
+    let maybe_entry_size = if max > fuzzy_pkg_size {
+        max
+    } else {
+        fuzzy_pkg_size
+    };
+    let mut buffer = String::with_capacity(MAYBE_CHAR_SIZE * maybe_entry_size);
 
-    async fn search_pkg(
-        &self,
-        pkg: &str,
-        max: usize,
-    ) -> anyhow::Result<(Option<String>, Vec<String>)> {
-        let exact_url = reqwest::Url::parse_with_params(SEARCH_BASE_URL, &[("name", pkg)])
-            .with_context(|| format!("{pkg} is a invalid params"))?;
-        let regex_url = reqwest::Url::parse_with_params(SEARCH_BASE_URL, &[("q", pkg)])
-            .with_context(|| format!("{pkg} is a invalid params"))?;
+    let mut push_into_buffer = |pkg: &ArchLinuxPkgInfo| {
+        let display = format!("<b>{}/{}</b>\n    {}", pkg.repo, pkg.pkgname, pkg.pkgdesc);
+        buffer.push_str(&display);
+        buffer.push('\n');
+    };
 
-        let (exact, regex) = tokio::join!(
-            self.to_t::<SearchResponse>(exact_url),
-            self.to_t::<SearchResponse>(regex_url),
-        );
-
-        let (exact, regex) = (exact?, regex?);
-
-        if !exact.valid || !regex.valid {
-            anyhow::bail!("invalid request!")
-        }
-
-        let results = regex
-            .results
-            .into_iter()
-            .take(max)
-            .map(|pkg| format!("<b>{}/{}</b>\n    {}", pkg.repo, pkg.pkgname, pkg.pkgdesc))
-            .collect();
-
-        if exact.results.is_empty() {
-            Ok((None, results))
-        } else {
-            let exact = &exact.results[0];
-            let exact = format!("{}/{}\n\t{}", exact.repo, exact.pkgname, exact.pkgdesc);
-            Ok((Some(exact), results))
-        }
+    if !exact_match.is_empty() {
+        push_into_buffer(&exact_match.results()[0])
     }
+
+    fuzzy_match
+        .results()
+        .iter()
+        .take(max)
+        .for_each(push_into_buffer);
+
+    Ok(Sendable::builder().text(buffer).build())
 }
