+++ src/modules/ehentai.rs
@@ -1,4 +1,77 @@
+use reqwest::Url;
 use serde::{Deserialize, Serialize};
+use std::iter::Iterator;
+
+use crate::app::AppData;
+
+use super::Sendable;
+
+pub async fn fetch_ehentai_comic_data<'a, I>(
+    data: AppData,
+    gid_list: I,
+) -> anyhow::Result<Vec<Sendable>>
+where
+    I: IntoIterator<Item = (&'a str, &'a str)>,
+{
+    let api_url: reqwest::Url = reqwest::Url::parse("https://api.e-hentai.org/api.php").unwrap();
+
+    let request_data = EhentaiRequestType::new(gid_list.into_iter());
+
+    let resp = data
+        .requester
+        .post_json_to_t::<PossibleEhentaiResponse>(&request_data, api_url)
+        .await?
+        .try_unwrap()?;
+
+    let v = resp.gmetadata.iter().fold(Vec::new(), |mut accum, elem| {
+        let display = format!(
+            "📖 Title: {}\
+             🗂️ Category: {}\
+             📄 Pages: {}\
+             ⭐ Rating: {}\
+             🔖 Tags: {}\
+             🔗 Links: {}\
+             🌱 Torrents ({}):\n{}
+            ",
+            elem.title_jpn,
+            elem.category,
+            elem.filecount,
+            elem.rating,
+            elem.tags.iter().fold(String::new(), |acc, x| format!(
+                "{acc} #{}",
+                x.split(':').nth(1).unwrap().replace([' ', '-'], "_")
+            )),
+            format_args!("https://e-hentai.org/g/{}/{}/", elem.gid, elem.token),
+            elem.torrentcount,
+            elem.torrents(std::cmp::max(3, elem.torrentcount) as usize)
+        );
+        let sendable = Sendable::Text(display);
+
+        accum.push(sendable);
+        accum
+    });
+
+    Ok(v)
+}
+
+pub fn parse_gid_list(s: &str) -> anyhow::Result<Vec<(&str, &str)>> {
+    let rules = regex::Regex::new(r#"e.hentai\.org/g/(\d+)/(\w+)"#).unwrap();
+    let capture: Vec<_> = rules
+        .captures_iter(s)
+        .filter_map(|cap| {
+            let a = cap.get(1)?.as_str();
+            let b = cap.get(2)?.as_str();
+
+            Some((a, b))
+        })
+        .collect();
+
+    if capture.is_empty() {
+        anyhow::bail!("No e-hentai or exhentai link found");
+    }
+
+    Ok(capture)
+}
 
 // -------------------------------- Type --------------------------------
 
@@ -6,19 +79,24 @@ use serde::{Deserialize, Serialize};
 /// will be genrated automatically.
 #[derive(Serialize, Debug)]
 pub struct EhentaiRequestType<'a> {
-    method: String,
+    method: &'static str,
     namespace: u8,
-    gidlist: &'a [[String; 2]],
+    gidlist: Vec<(&'a str, &'a str)>,
 }
 
 impl<'a> EhentaiRequestType<'a> {
     /// Require gid_list to consturct the request.
     /// Read https://ehwiki.org/wiki/API for details.
-    pub fn new(gidlist: &'a [[String; 2]]) -> Self {
+    pub fn new<I>(gidlist: I) -> Self
+    where
+        I: Iterator<Item = (&'a str, &'a str)>,
+    {
+        let gidlist = gidlist.collect::<Vec<_>>();
+
         Self {
             gidlist,
-            method: "gdata".to_string(),
             namespace: 1,
+            method: "gdata",
         }
     }
 }
@@ -30,7 +108,7 @@ macro_rules! gen_str_to_t {
     ($($ty:ty),+) => {
         paste::paste! {
             $(
-                fn [<to_$ty>]<'de, D>(d: D) -> Result<$ty, D::Error>
+                fn [<to_ $ty:lower >]<'de, D>(d: D) -> Result<$ty, D::Error>
                 where
                     D: serde::de::Deserializer<'de>,
                 {
@@ -43,17 +121,7 @@ macro_rules! gen_str_to_t {
     };
 }
 
-gen_str_to_t!(u32, u64, f32);
-
-/// Parse string to actual URL type
-fn to_url<'de, D>(d: D) -> Result<reqwest::Url, D::Error>
-where
-    D: serde::de::Deserializer<'de>,
-{
-    let orig: String = serde::de::Deserialize::deserialize(d)?;
-    use serde::de::Error;
-    reqwest::Url::parse(&orig).map_err(D::Error::custom)
-}
+gen_str_to_t![u32, u64, f32, Url];
 
 /// Represent data for single comic information
 #[derive(Deserialize, Debug)]
@@ -80,55 +148,31 @@ impl EhGmetadata {
     /// This function will render at most `max` torrent.
     /// If the `max` argument is larger then torrents amount,
     /// all the torrents information will be rendered.
-    pub fn to_telegram_html(&self, max: usize) -> String {
+    fn torrents(&self, max: usize) -> String {
         let gid = if let Some(ref id) = self.first_gid {
             id.to_string()
         } else {
             format!("{}", self.gid)
         };
 
-        self.torrents
+        let display = self
+            .torrents
             .iter()
             .take(max)
             .fold(String::new(), |sum, torrent| {
                 format!(
-                    r#"{sum}
-* <b>Name</b>: <a href="https://ehtracker.org/get/{}/{}.torrent">{}</a>
-  <b>Size</b>: {} MB
-"#,
+                    "{sum} \
+                     * <b>Name</b>: <a href=\"https://ehtracker.org/get/{}/{}.torrent\">{}</a> \
+                     * <b>Size</b>: {} MB
+                    ",
                     gid,
                     torrent.hash,
                     torrent.name,
                     torrent.fsize / 1000000,
                 )
-            })
-    }
-}
+            });
 
-/// Custom format to display comit data.
-impl std::fmt::Display for EhGmetadata {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            r#"📖 Title: {}
-🗂️ Category: {}
-📄 Pages: {}
-⭐ Rating: {}
-🌱 Torrent Amount: {}
-🔖 Tags: {}
-🔗 Links: {}
-"#,
-            self.title_jpn,
-            self.category,
-            self.filecount,
-            self.rating,
-            self.torrentcount,
-            self.tags.iter().fold(String::new(), |acc, x| format!(
-                "{acc} #{}",
-                x.split(':').nth(1).unwrap().replace([' ', '-'], "_")
-            )),
-            format_args!("https://e-hentai.org/g/{}/{}/", self.gid, self.token),
-        )
+        display
     }
 }
 
@@ -170,32 +214,5 @@ impl PossibleEhentaiResponse {
             Self::Norm(res) => Ok(res),
             Self::Err(e) => Err(anyhow::anyhow!("{}", e.error)),
         }
-    }
-}
-
-#[async_trait::async_trait]
-pub trait EhentaiProvider {
-    type Result;
-    async fn fetch_ehentai_comic_data(&self, gid_list: &[[String; 2]]) -> Self::Result;
-}
-
-#[async_trait::async_trait]
-impl EhentaiProvider for crate::maid::Fetcher {
-    type Result = anyhow::Result<EhentaiMetadataResponse>;
-
-    async fn fetch_ehentai_comic_data(&self, gid_list: &[[String; 2]]) -> Self::Result {
-        let api_url: reqwest::Url =
-            reqwest::Url::parse("https://api.e-hentai.org/api.php").unwrap();
-
-        let request_data = EhentaiRequestType::new(gid_list);
-
-        self.c
-            .post(api_url)
-            .json(&request_data)
-            .send()
-            .await?
-            .json::<PossibleEhentaiResponse>()
-            .await?
-            .try_unwrap()
     }
 }
