+++ src/modules/ksyx.rs
@@ -1,4 +1,6 @@
-#[async_trait::async_trait]
-pub trait KsyxCounterCache: Send + Sync + Clone {
-    async fn hit(&mut self) -> anyhow::Result<u32>;
+use crate::app::AppData;
+use redis::Commands;
+
+pub fn hit(data: AppData) -> anyhow::Result<u32> {
+    Ok(data.cacher.get_conn().incr("KSYX_HIT_COUNTER", 1)?)
 }
